backend/requirements.txt
# Core web framework
fastapi==0.110.1
uvicorn==0.25.0

# Database and auth
boto3>=1.34.129
requests-oauthlib>=2.0.0
cryptography>=42.0.8
python-dotenv>=1.0.1
pymongo==4.5.0
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
tzdata>=2024.2
motor==3.3.1
python-jose>=3.3.0
requests>=2.31.0
python-multipart>=0.0.9

# Development tools
pytest>=8.0.0
black>=24.1.1
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Core data processing
pandas>=2.2.0
numpy>=1.26.0,<2.0
jq>=1.6.0
typer>=0.9.0
bcrypt>=4.0.1
pillow>=10.0.0
pillow-heif>=0.13.0
aiofiles
psutil>=5.9.0

# 3D processing (core functionality)
mediapipe>=0.10.11
trimesh[easy]==4.0.5
open3d>=0.17.0
pybullet==3.2.7
opencv-python==4.8.1.78
scikit-image==0.22.0
scipy==1.11.4
scikit-learn==1.3.2

# External services
openai>=1.0.0
google-cloud-aiplatform>=1.38.0
celery>=5.3.0
redis>=5.0.0
flower>=2.0.1
aiohttp>=3.9.1

# ML stack - TESTED WORKING VERSIONS
torch>=2.8.0
torchvision>=0.19.0
torchaudio>=2.8.0
transformers>=4.45.0
diffusers>=0.30.0
huggingface-hub>=0.35.0
tokenizers>=0.22.0
accelerate>=0.25.0
safetensors>=0.4.1
